pyproject.toml
[tool.poetry]
name = "dummy"
version = "0.1.0"
description = ""
authors = ["imanmaris99 <imanmaris99@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "3.0.2"
flask-sqlalchemy = "3.1.1"
psycopg2 = "^2.9.9"
bcrypt = "4.1.2"
blinker = "1.7.0"
cachetools = "5.3.2"
cerberus = "1.3.5"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
docopt = "0.6.2"
flask-jwt-extended = "4.6.0"
flask-login = "0.6.3"
google-auth = "2.28.0"
idna = "3.6"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
markupsafe = "2.1.5"
mysql-connector-python = "8.3.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pyjwt = "2.8.0"
python-dotenv = "1.0.1"
requests = "2.31.0"
rsa = "4.9"
sqlalchemy = "2.0.25"
typing-extensions = "4.9.0"
urllib3 = "2.2.0"
werkzeug = "3.0.1"
yarg = "0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
